+++ medil/ecc_algorithms.py
@@ -24,46 +24,48 @@ def find_clique_min_cover(graph, verbose=True):
             max_intersect_num = p + t if p > 0 else 1
         print("solution has at most {} cliques.".format(max_intersect_num))
     while the_cover is None:
-        reducible_graph = graph.reducible_copy()
         if True:                # verbose:
             print("\ntesting for solutions with {}/{} cliques".format(num_cliques, max_intersect_num))
-        the_cover = branch(reducible_graph, num_cliques, the_cover)
+        the_cover = branch(graph, num_cliques, the_cover)
         num_cliques += 1
 
     return reducible_graph.reconstruct_cover(the_cover) # according to rule_3
 
 
-def branch(reducible_graph, k_num_cliques, the_cover):
+def branch(graph, k_num_cliques, the_cover):
+    branch_graph = graph.reducible_copy()
     if the_cover is not None:
-        for clique in the_cover:
-            reducible_graph.the_cover = clique
-            reducible_graph.cover_edges()  # only works one clique at a time, or on a list of edges
-    if reducible_graph.num_edges == 0:
+        print(the_cover)
+        for clique in the_cover:  # this might not be necessary, since the_cover_prime is only +1 clique
+            print('clique: {}'.format{clique})
+            branch_graph.the_cover = [clique]
+            branch_graph.cover_edges()  # only works one clique at a time, or on a list of edges
+    if branch_graph.num_edges == 0:
         return the_cover
     else:
-        reducible_graph.the_cover = the_cover
+        branch_graph.the_cover = the_cover
 
-    if reducible_graph.verbose:
+    if branch_graph.verbose:
         print("\tbranching...")
 
-    reducible_graph.reduzieren(k_num_cliques)
-    k_num_cliques = reducible_graph.k_num_cliques
+    branch_graph.reduzieren(k_num_cliques)
+    k_num_cliques = branch_graph.k_num_cliques
     
     if k_num_cliques < 0:
         return None
 
-    if reducible_graph.num_edges == 0:
-        return reducible_graph.the_cover  # not in paper, but seems necessary?
+    if branch_graph.num_edges == 0:
+        return graph.the_cover  # not in paper, but speeds it up slightly
 
-    chosen_nbrhood = reducible_graph.choose_nbrhood()
+    chosen_nbrhood = branch_graph.choose_nbrhood()
     for clique_nodes in max_cliques(chosen_nbrhood):
         if len(clique_nodes) == 1:  # then this vert has been rmed; quirk of max_cliques
             continue
-        clique = np.zeros(reducible_graph.unreduced.num_vertices, dtype=int)
+        clique = np.zeros(branch_graph.unreduced.num_vertices, dtype=int)
         clique[clique_nodes] = 1
-        union = clique.reshape(1, -1) if reducible_graph.the_cover is None else np.vstack((reducible_graph.the_cover, clique))
+        union = clique.reshape(1, -1) if branch_graph.the_cover is None else np.vstack((branch_graph.the_cover, clique))
 
-        the_cover_prime = branch(reducible_graph, k_num_cliques-1, union)
+        the_cover_prime = branch(branch_graph, k_num_cliques-1, union)
         if the_cover_prime is not None:
             return the_cover_prime
     return None