+++ CHANGELOG.md
@@ -18,13 +18,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed ###
 - bug in max\_intersection\_num calculation
 - exception handling for edgeless graph
-- bug with graph.cover_edges() use in ecc.branch()
+- bug with graph.cover\_edges() use in ecc.find\_cm.branch()
+- bug with using graph copy outside of ecc.find\_cm.branch() rather than within
 
 ### Removed ###
-- hack to stop while loop in find_clique_min_cover()
+- hack to stop while loop in find\_clique\_min\_cover()
 
 ### Changed ###
 - default verbosity of output
+- reducible\_graph var name to branch\_grap
 
 ## [0.3.0] - 2020-04-13 ##
 ### Added ###